requirements.txt
google-generativeai
PyPDF2
numpy
faiss-cpu
streamlit
python-dotenv 